refactor(api): split task and search facades
Move task dispatch, task status, search filters, semantic helpers, and trend helpers behind stable facade modules. Keep existing import and monkeypatch seams live while reducing large runtime files.
