+++ pipeline/task_summary_generation.py
@@ -2,11 +2,9 @@ from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Any
 
-from pipeline.agenda_summary_maintenance import AGENDA_SUMMARY_EMBED_DISPATCH_ERRORS
 from pipeline.llm import LocalAI
 from pipeline.models import AgendaItem, Catalog, Document
-from pipeline.task_runtime import logger
-from pipeline.task_side_effects import REINDEX_FAILURE_EXCEPTIONS
+from pipeline.task_summary_side_effects import run_summary_generation_side_effects
 
 
 AGENDA_DOC_KIND = "agenda"
@@ -36,41 +34,6 @@ class SummaryGenerationTaskServices:
     persist_agenda_summary: Callable[..., dict[str, Any]]
     reindex_catalog: Callable[[int], object]
     embed_catalog: Callable[[int], object]
-
-
-def run_summary_generation_side_effects(
-    catalog_id: int,
-    *,
-    services: SummaryGenerationTaskServices,
-) -> dict[str, int]:
-    """
-    Summary persistence is the source of truth; search and embedding side effects stay best-effort.
-    """
-    reindexed = 0
-    reindex_failed = 0
-    try:
-        services.reindex_catalog(catalog_id)
-        reindexed = 1
-    except REINDEX_FAILURE_EXCEPTIONS as reindex_error:
-        reindex_failed = 1
-        logger.warning("summary_generation.reindex_failed catalog_id=%s error=%s", catalog_id, reindex_error)
-
-    embed_enqueued = 0
-    embed_dispatch_failed = 0
-    try:
-        services.embed_catalog(catalog_id)
-        embed_enqueued = 1
-    except AGENDA_SUMMARY_EMBED_DISPATCH_ERRORS as dispatch_error:
-        logger.warning("embed_catalog_task.dispatch_failed catalog_id=%s error=%s", catalog_id, dispatch_error)
-        embed_dispatch_failed = 1
-
-    return {
-        "reindexed": reindexed,
-        "reindex_failed": reindex_failed,
-        "embed_enqueued": embed_enqueued,
-        "embed_dispatch_failed": embed_dispatch_failed,
-    }
-
 
 def _source_text_quality_payload(catalog: Catalog, services: SummaryGenerationTaskServices) -> dict[str, Any] | None:
     if not catalog.content:
